📝 Add `docs` script target (using typedoc)
📝 Add documentation for several functions
♻ Move most types into src/types.ts
This is temporary until we can better organize the project, as a
refactor is needed prior to 0.2 so that we can more effectively filter
out releases users don't want.
♻ After renaming several types, we also made most of the
filter/match/seeker types generic.

The following is now a valid way to create a Seeker<T>:
```
const m = matcher<Release>(/v1.*/);
```
Note that `matcher` will be renamed to `seeker` to be more correct