➕ Add dotenv as dev dependency
This is used for loading a local `.env` with a GITHUB_TOKEN PAT set.
This is now the only way to run some tests locally. Simply use a
public_repo scoped token.

✅ Improve coverage by cleaning up branches
✅ Improve coverage by removing unauthenticated clients
🐛 Fix some incorrectly created interface members (URL vs string)
♻ Simplify the interface for ReleaseOptions and the call to releases.
We will need to support calling maxSatisfying in some manner for basic
usage, but the responsibility of finding a single release is no longer a
responsibility of the library
💡 Update comments regarding TODOs
✨ Export semver.validRange by default